+++ core/urls.py
@@ -1,14 +1,9 @@
-# -*- encoding: utf-8 -*-
-"""
-Copyright (c) 2019 - present AppSeed.us
-"""
-
 from django.contrib import admin
-from django.urls import path, include
+from django.urls import include, path
 
 try:
     from rest_framework.authtoken.views import obtain_auth_token
-except:
+except Exception:
     pass
 
 urlpatterns = [
@@ -23,5 +18,5 @@ urlpatterns = [
 try:
     urlpatterns.append(path("api/", include("api.urls")))
     urlpatterns.append(path("login/jwt/", view=obtain_auth_token))
-except:
+except Exception:
     pass
