+++ home/models.py
@@ -6,6 +6,7 @@ from django.contrib.auth.models import User
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
+from helpers.util import format_currency
 from home.consts import (
     CASHIER,
     CURRENCY_CHOICES,
@@ -14,7 +15,6 @@ from home.consts import (
     ROLE_CHOICES,
     STATUS_CHOICES,
 )
-from home.utils import format_currency
 
 
 def create_object_id():
