+++ home/views.py
@@ -1,14 +1,7 @@
-# -*- encoding: utf-8 -*-
-"""
-Copyright (c) 2019 - present AppSeed.us
-"""
-
 from django.shortcuts import render
-from django.http import HttpResponse
 
-# Create your views here.
 
 def index(request):
 
-    # Page from the theme 
-    return render(request, 'pages/index.html')
+    # Page from the theme
+    return render(request, "pages/index.html")
